refactor: break down into components
