+++ src/types/types.ts
@@ -4,3 +4,9 @@ export type ThemeContextType = {
   themeValue: ThemeValue;
   setThemeValue: (theme: ThemeValue) => void;
 };
+
+export type VariablesData = {
+  key: number;
+  variable: string;
+  value: string;
+};
